Initial support for Quality of Service
This spec intends to define an initial way/PoC of having QoS in GBP.

This PoC will implement working QoS support for the scenario where QoS rules
are configured for Policy Targets. However, it will be done by inheriting from
a new kind of Network Service Policy (NSP) that will be assigned to a PTG.

The NSP resource mapping logic will proceed to create the corresponding QoS
policies/rules in Neutron to reflect what is specified in the NSPs, and then
map these to the Policy Targets that needs to share the QoS policies/rules.

This change will be attempted in a specific feature branch based on GBP master
supporting upstream Neutron Liberty integration.

Change-Id: Ie7e5f52e4fa91dab98b22f8310fb23750af04405