Grouping of PTGs based on Application Characteristics
A Policy Target Group (PTG) can be used to represent a specific tier of an
Application (e.g. DB tier). However, there is no construct in GBP that
corresponds to the application itself. Such a construct can help to abstract
the properties of the application (across the PTGs) futher simplifying the
intent specification and automated orchestration.

It is proposed to add a new resource Application Policy Group (APG) that will
have a 1 to many relationship with PTGs. Each PTG will belong to exactly one
APG. The APG does not have relationship (or constraints) with any other GBP
resources.

Change-Id: I45fd722ccd000ff311b37088f0d96a8a3411eaaa
Defines: blueprint application-group