L3-Policy mapping to Address Scope & Subnetpool
Starting in the Mitaka release Neutron defines and supports the use of Address
Scopes. Subnet-pools were defined in the earlier Kilo release and can now be
used within Address Scopes. The notion of an address scope has existed in GBP
in the form of L3-Policy. The GBP L3-Policy manages the address-scope and
subnet-pools with its own home-grown implementation. It would be preferable to
leverage the Neutron address-scopes and subnet-pools instead.

Change-Id: I9e4335073dc8ec0cb85db909560c49e0dddefe82
Defines: blueprint address-scope-mapping